+++ packages/vmind/src/gpt/chart-generation/index.ts
@@ -1,1 +1,1 @@
-export * from './NLToChartPipe';
+export * from './NLToChart';
